Delete entire module working

+++ src/main/java/seedu/command/DeleteCommand.java
@@ -12,8 +12,8 @@ import java.time.LocalDate;
  */
 public class DeleteCommand extends Command {
     private int taskNum;
-    private String type;
-    private String description;
+    private String type= null;
+    private String description = null;
     private LocalDate date = null;
     private String moduleCode = null;
 
@@ -21,19 +21,19 @@ public class DeleteCommand extends Command {
         this.description = description;
     }
 
-    public DeleteCommand(String description, String type) {
+    public DeleteCommand(String moduleCode, String type) {
+        this.moduleCode = moduleCode.strip();
         this.type = type;
-        this.description = description;
     }
 
     public DeleteCommand(String description, LocalDate date) {
-        this.description = description;
+        this.description = description.strip();
         this.date = date;
     }
 
 
     public DeleteCommand(String moduleCode, LocalDate date, String description) {
-        this.description = description;
+        this.description = description.strip();
         this.moduleCode = moduleCode;
         this.date = date;
     }
@@ -45,20 +45,32 @@ public class DeleteCommand extends Command {
 
     @Override
     public void execute(ScheduleManager scheduleManager, ModuleManager moduleManager, Ui ui) {
-        description = description.strip();
-        if(this.date == null){
-            if(moduleCode.equals(null)){
-//                moduleManager.delete();
+        System.out.println(moduleCode);
+        if(moduleCode.equals(null)){
+            if(this.date == null){                 //delete task with no date
+                System.out.println("delete task with no date");
+                scheduleManager.deleteTask(description);
+                return;
             }
-            scheduleManager.deleteTask(description);
-            moduleManager.deleteTask(description);
-            Storage.getStorage().exportData(moduleManager);
+            System.out.println("delete task with date, with description");
+            scheduleManager.deleteTask(description, date);//delete task with date, with description
+            return;
+        }
+        if(this.date == null){
+            System.out.println("delete entire module");
+            moduleManager.delete(moduleCode);
+            scheduleManager.deleteTask("module",moduleCode);            //delete entire module
         }
         else{
-            scheduleManager.deleteTask(description, date);
-            moduleManager.deleteTask(description, date);
-            String commandString = "delete " + description + " /date " + this.date;
-            Storage.getStorage().exportAdditionalData(commandString);
+            if(description.equals(null)){                           //delete all task in module matching date
+                System.out.println("delete all task in module matching date");
+                moduleManager.delete(moduleCode,date);
+                scheduleManager.deleteTask(description, date);
+                return;
+            }System.out.println("delete all task in module with matching date and description");
+            moduleManager.delete(moduleCode,description,date);         //delete all task in module with matching date and description
+            scheduleManager.deleteTask(moduleCode, description, date);
+
         }
     }
 }
